requirements.txt
asgiref==3.7.2
certifi==2024.2.2
channels==4.0.0
charset-normalizer==3.3.2
crispy-bootstrap4==2024.1
dj-database-url==2.1.0
Django==5.0.2
django-crispy-forms==2.1
idna==3.6
paystackapi==2.1.1
pillow==10.2.0
psycopg2==2.9.9
python-decouple==3.8
requests==2.31.0
setuptools==69.1.1
simplejson==3.19.2
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.10.0
tzdata==2024.1
urllib3==2.2.1
zope.interface==6.2
